tests 11 and 13

+++ src/user/tests/tlib/sysapi.h
@@ -132,7 +132,7 @@ void sys_info(void);
 typedef __SIZE_TYPE__ size_t;
 
 int strcmp(const char *str1, const char *str2);
-unsigned long strlen(const char *s);
+size_t strlen(const char *s);
 char *strncpy(char *, const char *, size_t);
 void *memset(void *dst, int c, size_t n);
